More Verbose Logging & Addition of Gitignore File
The server's connection message now logs the port the server is connected to as well as the date and time it connected.

Added a (somewhat redundant) getDate() function to get a brand new Date object and access it's methods.

Added the basic framework for choosing a room in chat.html.

Added a .gitignore file to prevent the node_modules from syncing with the repo.