Server.js Accounts Update
A major update to server.js that more than doubles the amount of code from the last version.

Added an encryption key and a cookie secret that are stored in a .env file and are accessed via the process environment variables.

Authenticate() will validate user credentials with a username and either a password or a unique 25-character-long token assigned at the time of account creation.

updateip() updates the list of safe ip addresses and their expiration dates for a specific account.

Added the useraccount class to streamline the creation of user accounts.

Added router for default path to direct clients to the proper login or chat page as needed.

The server will check request bodies made to /newacc against several regular expressions to validate user-supplied information, check for duplicate usernames or emails, and create new user accounts with valid user info.

After a request to /authenticate successfully completes the auth challenge, the server will send cookies to the client to enable login persistence.

The server now derives the message author directly from the cookies contained in the socket handshake for members and labels usernames provided by guest users.